fix: improve package structure
