+++ app/src/main/kotlin/com/softteco/template/ui/feature/onboarding/login/LoginViewModel.kt
@@ -1,14 +1,14 @@
-package com.softteco.template.ui.feature.forgotPassword
+package com.softteco.template.ui.feature.onboarding.login
 
 import androidx.compose.runtime.Immutable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.softteco.template.R
 import com.softteco.template.data.base.error.AppError.AuthError.EmailNotExist
-import com.softteco.template.data.base.error.AppError.AuthError.InvalidEmail
+import com.softteco.template.data.base.error.AppError.AuthError.WrongCredentials
 import com.softteco.template.data.base.error.Result
 import com.softteco.template.data.profile.ProfileRepository
-import com.softteco.template.data.profile.dto.ResetPasswordDto
+import com.softteco.template.data.profile.dto.CredentialsDto
 import com.softteco.template.navigation.Screen
 import com.softteco.template.ui.components.FieldState
 import com.softteco.template.ui.components.FieldType
@@ -24,26 +24,26 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class ForgotPasswordViewModel @Inject constructor(
+class LoginViewModel @Inject constructor(
     private val repository: ProfileRepository,
     private val appDispatchers: AppDispatchers,
     private val snackbarController: SnackbarController,
     private val dialogController: DialogController,
 ) : ViewModel() {
+
     private val loading = MutableStateFlow(false)
     private var emailState = MutableStateFlow(TextFieldState())
-    private val ctaButtonState = emailState.map { email ->
-        email.text.validateInputValue(FieldType.EMAIL) is FieldState.Valid
-    }
+    private var passwordState = MutableStateFlow(TextFieldState())
+    private val ctaButtonState = MutableStateFlow(false)
 
     private val _navDestination = MutableSharedFlow<Screen>(
         extraBufferCapacity = 1,
@@ -54,17 +54,23 @@ class ForgotPasswordViewModel @Inject constructor(
     val state = combine(
         loading,
         emailState,
+        passwordState,
         ctaButtonState,
-    ) { loading, email, isCtaEnabled ->
+    ) { loading, email, password, isCtaEnabled ->
         State(
             loading = loading,
             email = email,
+            password = password,
             onEmailChanged = {
                 emailState.value = TextFieldState(it, FieldState.AwaitingInput)
             },
-            onInputComplete = ::onInputComplete,
-            isResetBtnEnabled = !loading && isCtaEnabled,
-            onRestorePasswordClicked = ::onForgotPassword,
+            onPasswordChanged = {
+                passwordState.value = TextFieldState(it, FieldState.AwaitingInput)
+            },
+            isLoginBtnEnabled = !loading && isCtaEnabled,
+            onInputComplete = { onInputComplete(it) },
+            onLoginClicked = ::onLogin,
+            onNavClick = { _navDestination.tryEmit(it) }
         )
     }.stateIn(
         viewModelScope,
@@ -72,23 +78,33 @@ class ForgotPasswordViewModel @Inject constructor(
         State()
     )
 
-    private fun onForgotPassword() {
-        viewModelScope.launch(appDispatchers.ui) {
+    init {
+        viewModelScope.launch {
+            combine(emailState, passwordState) { email, password ->
+                ctaButtonState.value =
+                    email.text.validateInputValue(FieldType.EMAIL) is FieldState.Valid &&
+                    password.text.validateInputValue(FieldType.PASSWORD) is FieldState.Valid
+            }.collect()
+        }
+    }
+
+    private fun onLogin() {
+        viewModelScope.launch(appDispatchers.io) {
             loading.value = true
 
-            val email = ResetPasswordDto(email = emailState.value.text)
+            val credentials = CredentialsDto(
+                email = emailState.value.text,
+                password = passwordState.value.text,
+            )
 
-            val result = repository.resetPassword(email)
+            val result = repository.login(credentials)
 
             when (result) {
-                is Result.Success -> {
-                    snackbarController.showSnackbar(R.string.check_email)
-                    _navDestination.tryEmit(Screen.Login)
-                }
+                is Result.Success -> _navDestination.tryEmit(Screen.Home)
 
                 is Result.Error -> {
                     when (result.error) {
-                        EmailNotExist, InvalidEmail -> {
+                        EmailNotExist, WrongCredentials -> {
                             emailState.update {
                                 TextFieldState(
                                     it.text,
@@ -105,10 +121,9 @@ class ForgotPasswordViewModel @Inject constructor(
 
                         else -> snackbarController.showSnackbar(result.error.messageRes)
                     }
+                    loading.value = false
                 }
             }
-
-            loading.value = false
         }
     }
 
@@ -127,17 +142,31 @@ class ForgotPasswordViewModel @Inject constructor(
         }
     }
 
-    private fun onInputComplete() {
-        emailState.update { TextFieldState(it.text, it.text.validateInputValue(FieldType.EMAIL)) }
+    private fun onInputComplete(fieldType: FieldType) {
+        when (fieldType) {
+            FieldType.EMAIL -> emailState.update {
+                TextFieldState(it.text, it.text.validateInputValue(fieldType))
+            }
+
+            FieldType.PASSWORD -> passwordState.update {
+                TextFieldState(it.text, it.text.validateInputValue(fieldType))
+            }
+
+            FieldType.USERNAME -> { /*NOOP*/
+            }
+        }
     }
 
     @Immutable
     data class State(
         val loading: Boolean = false,
         val email: TextFieldState = TextFieldState(),
+        val password: TextFieldState = TextFieldState(),
         val onEmailChanged: (String) -> Unit = {},
-        val onInputComplete: () -> Unit = {},
-        val isResetBtnEnabled: Boolean = false,
-        val onRestorePasswordClicked: () -> Unit = {},
+        val onPasswordChanged: (String) -> Unit = {},
+        val onInputComplete: (FieldType) -> Unit = {},
+        val isLoginBtnEnabled: Boolean = false,
+        val onLoginClicked: () -> Unit = {},
+        val onNavClick: (Screen) -> Unit = {},
     )
 }
