+++ app/src/main/kotlin/com/softteco/template/ui/feature/onboarding/password/forgot/ForgotPasswordScreen.kt
@@ -1,14 +1,14 @@
-package com.softteco.template.ui.feature.resetPassword
+package com.softteco.template.ui.feature.onboarding.password.forgot
 
 import android.content.res.Configuration
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -23,11 +23,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.softteco.template.R
 import com.softteco.template.navigation.Screen
-import com.softteco.template.ui.components.PasswordField
+import com.softteco.template.ui.components.AppTextField
+import com.softteco.template.ui.components.CustomTopAppBar
 import com.softteco.template.ui.components.PrimaryButton
 import com.softteco.template.ui.theme.AppTheme
 import com.softteco.template.ui.theme.Dimens
-import com.softteco.template.ui.theme.Dimens.PaddingExtraLarge
 import com.softteco.template.utils.Analytics
 import com.softteco.template.utils.LockScreenOrientation
 import kotlinx.coroutines.flow.launchIn
@@ -35,72 +35,89 @@ import kotlinx.coroutines.flow.onEach
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun ResetPasswordScreen(
+fun ForgotPasswordScreen(
+    onBackClicked: () -> Unit,
     onSuccess: () -> Unit,
+    navigateToSignUp: () -> Unit,
     modifier: Modifier = Modifier,
-    viewModel: ResetPasswordViewModel = hiltViewModel()
+    viewModel: ForgotPasswordViewModel = hiltViewModel(),
 ) {
     LockScreenOrientation(Configuration.ORIENTATION_PORTRAIT)
 
     val state by viewModel.state.collectAsState()
 
     LaunchedEffect(Unit) {
-        Analytics.resetPasswordOpened()
+        Analytics.forgotPasswordOpened()
 
         viewModel.navDestination.onEach { screen ->
-            if (screen == Screen.Login) {
-                Analytics.resetPasswordSuccess()
-                onSuccess()
+            when (screen) {
+                Screen.Login -> onSuccess()
+                Screen.SignUp -> navigateToSignUp()
+
+                else -> { /*NOOP*/
+                }
             }
         }.launchIn(this)
     }
 
     val keyboardController = LocalSoftwareKeyboardController.current
     ScreenContent(
-        state = state,
         modifier = modifier.pointerInput(Unit) {
             detectTapGestures(onTap = { keyboardController?.hide() })
         },
+        state = state,
+        onBackClicked = onBackClicked
     )
 }
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
 private fun ScreenContent(
-    state: ResetPasswordViewModel.State,
+    state: ForgotPasswordViewModel.State,
     modifier: Modifier = Modifier,
+    onBackClicked: () -> Unit,
 ) {
     Column(
         modifier = modifier
             .background(MaterialTheme.colorScheme.background)
-            .padding(PaddingExtraLarge)
             .fillMaxSize(),
+        verticalArrangement = Arrangement.spacedBy(Dimens.PaddingExtraLarge),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Text(text = stringResource(id = R.string.enter_new_password))
-        PasswordField(
-            passwordValue = state.password.text,
-            onPasswordChanged = state.onPasswordChanged,
-            fieldStatePassword = state.password.state,
-            onInputComplete = state.onInputComplete,
-            modifier = Modifier
-                .padding(top = PaddingExtraLarge)
-                .fillMaxWidth()
+        CustomTopAppBar(
+            stringResource(id = R.string.forgot_password),
+            showBackIcon = true,
+            modifier = Modifier.fillMaxWidth(),
+            onBackClicked = onBackClicked
         )
+        Column(
+            modifier = Modifier.padding(Dimens.PaddingNormal),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            AppTextField(
+                value = state.email.text,
+                onValueChanged = state.onEmailChanged,
+                fieldState = state.email.state,
+                onInputComplete = state.onInputComplete,
+                labelRes = R.string.email,
+                modifier = Modifier.fillMaxWidth()
+            )
 
-        val keyboardController = LocalSoftwareKeyboardController.current
-        PrimaryButton(
-            buttonText = stringResource(id = R.string.reset_password),
-            loading = state.loading,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = Dimens.PaddingDefault),
-            enabled = state.isResetBtnEnabled,
-            onClick = {
-                state.onResetPasswordClicked()
-                keyboardController?.hide()
-            }
-        )
+            val keyboardController = LocalSoftwareKeyboardController.current
+            PrimaryButton(
+                buttonText = stringResource(id = R.string.restore_password),
+                loading = state.loading,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = Dimens.PaddingLarge),
+                enabled = state.isResetBtnEnabled,
+                onClick = {
+                    state.onRestorePasswordClicked()
+                    keyboardController?.hide()
+                }
+            )
+        }
     }
 }
 
@@ -108,8 +125,6 @@ private fun ScreenContent(
 @Composable
 private fun Preview() {
     AppTheme {
-        ScreenContent(
-            ResetPasswordViewModel.State()
-        )
+        ScreenContent(state = ForgotPasswordViewModel.State(), onBackClicked = {})
     }
 }
