+++ app/src/main/kotlin/com/softteco/template/ui/feature/onboarding/password/forgot/ForgotPasswordViewModel.kt
@@ -1,14 +1,14 @@
-package com.softteco.template.ui.feature.login
+package com.softteco.template.ui.feature.onboarding.password.forgot
 
 import androidx.compose.runtime.Immutable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.softteco.template.R
 import com.softteco.template.data.base.error.AppError.AuthError.EmailNotExist
-import com.softteco.template.data.base.error.AppError.AuthError.WrongCredentials
+import com.softteco.template.data.base.error.AppError.AuthError.InvalidEmail
 import com.softteco.template.data.base.error.Result
 import com.softteco.template.data.profile.ProfileRepository
-import com.softteco.template.data.profile.dto.CredentialsDto
+import com.softteco.template.data.profile.dto.ResetPasswordDto
 import com.softteco.template.navigation.Screen
 import com.softteco.template.ui.components.FieldState
 import com.softteco.template.ui.components.FieldType
@@ -24,26 +24,26 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class LoginViewModel @Inject constructor(
+class ForgotPasswordViewModel @Inject constructor(
     private val repository: ProfileRepository,
     private val appDispatchers: AppDispatchers,
     private val snackbarController: SnackbarController,
     private val dialogController: DialogController,
 ) : ViewModel() {
-
     private val loading = MutableStateFlow(false)
     private var emailState = MutableStateFlow(TextFieldState())
-    private var passwordState = MutableStateFlow(TextFieldState())
-    private val ctaButtonState = MutableStateFlow(false)
+    private val ctaButtonState = emailState.map { email ->
+        email.text.validateInputValue(FieldType.EMAIL) is FieldState.Valid
+    }
 
     private val _navDestination = MutableSharedFlow<Screen>(
         extraBufferCapacity = 1,
@@ -54,23 +54,17 @@ class LoginViewModel @Inject constructor(
     val state = combine(
         loading,
         emailState,
-        passwordState,
         ctaButtonState,
-    ) { loading, email, password, isCtaEnabled ->
+    ) { loading, email, isCtaEnabled ->
         State(
             loading = loading,
             email = email,
-            password = password,
             onEmailChanged = {
                 emailState.value = TextFieldState(it, FieldState.AwaitingInput)
             },
-            onPasswordChanged = {
-                passwordState.value = TextFieldState(it, FieldState.AwaitingInput)
-            },
-            isLoginBtnEnabled = !loading && isCtaEnabled,
-            onInputComplete = { onInputComplete(it) },
-            onLoginClicked = ::onLogin,
-            onNavClick = { _navDestination.tryEmit(it) }
+            onInputComplete = ::onInputComplete,
+            isResetBtnEnabled = !loading && isCtaEnabled,
+            onRestorePasswordClicked = ::onForgotPassword,
         )
     }.stateIn(
         viewModelScope,
@@ -78,33 +72,23 @@ class LoginViewModel @Inject constructor(
         State()
     )
 
-    init {
-        viewModelScope.launch {
-            combine(emailState, passwordState) { email, password ->
-                ctaButtonState.value =
-                    email.text.validateInputValue(FieldType.EMAIL) is FieldState.Valid &&
-                    password.text.validateInputValue(FieldType.PASSWORD) is FieldState.Valid
-            }.collect()
-        }
-    }
-
-    private fun onLogin() {
-        viewModelScope.launch(appDispatchers.io) {
+    private fun onForgotPassword() {
+        viewModelScope.launch(appDispatchers.ui) {
             loading.value = true
 
-            val credentials = CredentialsDto(
-                email = emailState.value.text,
-                password = passwordState.value.text,
-            )
+            val email = ResetPasswordDto(email = emailState.value.text)
 
-            val result = repository.login(credentials)
+            val result = repository.resetPassword(email)
 
             when (result) {
-                is Result.Success -> _navDestination.tryEmit(Screen.Home)
+                is Result.Success -> {
+                    snackbarController.showSnackbar(R.string.check_email)
+                    _navDestination.tryEmit(Screen.Login)
+                }
 
                 is Result.Error -> {
                     when (result.error) {
-                        EmailNotExist, WrongCredentials -> {
+                        EmailNotExist, InvalidEmail -> {
                             emailState.update {
                                 TextFieldState(
                                     it.text,
@@ -121,9 +105,10 @@ class LoginViewModel @Inject constructor(
 
                         else -> snackbarController.showSnackbar(result.error.messageRes)
                     }
-                    loading.value = false
                 }
             }
+
+            loading.value = false
         }
     }
 
@@ -142,31 +127,17 @@ class LoginViewModel @Inject constructor(
         }
     }
 
-    private fun onInputComplete(fieldType: FieldType) {
-        when (fieldType) {
-            FieldType.EMAIL -> emailState.update {
-                TextFieldState(it.text, it.text.validateInputValue(fieldType))
-            }
-
-            FieldType.PASSWORD -> passwordState.update {
-                TextFieldState(it.text, it.text.validateInputValue(fieldType))
-            }
-
-            FieldType.USERNAME -> { /*NOOP*/
-            }
-        }
+    private fun onInputComplete() {
+        emailState.update { TextFieldState(it.text, it.text.validateInputValue(FieldType.EMAIL)) }
     }
 
     @Immutable
     data class State(
         val loading: Boolean = false,
         val email: TextFieldState = TextFieldState(),
-        val password: TextFieldState = TextFieldState(),
         val onEmailChanged: (String) -> Unit = {},
-        val onPasswordChanged: (String) -> Unit = {},
-        val onInputComplete: (FieldType) -> Unit = {},
-        val isLoginBtnEnabled: Boolean = false,
-        val onLoginClicked: () -> Unit = {},
-        val onNavClick: (Screen) -> Unit = {},
+        val onInputComplete: () -> Unit = {},
+        val isResetBtnEnabled: Boolean = false,
+        val onRestorePasswordClicked: () -> Unit = {},
     )
 }
